Digitalis: TBI mask, ASR Wd zero-ext, UBFM general JIT, USHL UB fix
Four independent correctness fixes uncovered while investigating an
ARM64 decompressor that consumed but mis-processed input on x86_64:

* TBI (Top Byte Ignore) for guest memory accesses. ARM64 ignores the
  top byte of pointers in load/store; x86_64 does not. Tagged guest
  pointers therefore faulted on the host. Mask via SHL 8 / SHR 8 of
  the address register before using it as a memory operand in Load,
  Store, SimdLoadStoreImm, SimdLoadStorePair, SimdLoadStoreReg, and
  LoadStoreExclusive. The kTopByteIgnore config flag enum existed in
  config_globals but was previously not wired into any code path.

* ASR Wd zero-extend. SBFM imms=31 32-bit form (ASR Wd, Wn, #imm)
  previously did Sarl + Movsxlq, sign-extending bit 31 of the shifted
  W-value into the upper 32 of Xd. ARM64 W-register writes always
  zero the upper 32 of Xd, so the Movsxlq is wrong: Sarl already
  zero-extends per x86 rules. Drop the Movsxlq.

* UBFM general handler in JIT. Previously only LSR / LSL / UXTB /
  UXTH aliases were JIT-emitted, leaving UBFX and UBFIZ to fall back
  to the interpreter. The decompressor's bit-stream reader executes
  UBFIZ on its hot path; tracing showed millions of interpreter
  bailouts per decompression run. Emit shift+mask directly for the
  general extract (imms >= immr) and insert (imms < immr) cases.

* INT8_MIN UB in USHL / SSHL / saturating-shift fallbacks. The
  per-lane shift value was loaded as int8_t then negated; `-(-128)`
  overflows int8_t and on x86 left rshift = -128, which the host's
  shift instruction reads modulo 64 = 0, returning `a` instead of
  the spec-correct 0. Promote shift and rshift to int32_t / uint32_t.

All 27 sample modules continue to PASS after these changes.

+++ interpreter/arm64/interpreter.h
@@ -2505,14 +2505,15 @@ class Interpreter {
       case Decoder::AdvSimdThreeSameOpcode::kSshl:
         AdvSimdThreeSameElementWise(src_n, src_m, esize, num_elements, &result,
             [](uint64_t a, uint64_t b, uint8_t es) -> uint64_t {
-              int8_t shift = static_cast<int8_t>(b & 0xFF);
-              uint8_t bits = es * 8;
+              // Promote to int32_t so -shift never overflows (int8_t INT8_MIN case).
+              int32_t shift = static_cast<int8_t>(b & 0xFF);
+              uint32_t bits = es * 8;
               if (shift >= 0) {
-                return (shift >= bits) ? 0 : (a << shift);
+                return (static_cast<uint32_t>(shift) >= bits) ? 0 : (a << shift);
               } else {
                 // Signed shift right: sign-extend a, then shift.
                 int64_t sa = static_cast<int64_t>(a << (64 - bits)) >> (64 - bits);
-                int8_t rshift = -shift;
+                uint32_t rshift = static_cast<uint32_t>(-shift);
                 return static_cast<uint64_t>(
                     (rshift >= bits) ? (sa >> (bits - 1)) : (sa >> rshift));
               }
@@ -2521,12 +2522,12 @@ class Interpreter {
       case Decoder::AdvSimdThreeSameOpcode::kUshl:
         AdvSimdThreeSameElementWise(src_n, src_m, esize, num_elements, &result,
             [](uint64_t a, uint64_t b, uint8_t es) -> uint64_t {
-              int8_t shift = static_cast<int8_t>(b & 0xFF);
-              uint8_t bits = es * 8;
+              int32_t shift = static_cast<int8_t>(b & 0xFF);
+              uint32_t bits = es * 8;
               if (shift >= 0) {
-                return (shift >= bits) ? 0 : (a << shift);
+                return (static_cast<uint32_t>(shift) >= bits) ? 0 : (a << shift);
               } else {
-                int8_t rshift = -shift;
+                uint32_t rshift = static_cast<uint32_t>(-shift);
                 return (rshift >= bits) ? 0 : (a >> rshift);
               }
             });
@@ -2542,12 +2543,12 @@ class Interpreter {
         // Fallback: treat as SSHL/USHL for basic functionality.
         AdvSimdThreeSameElementWise(src_n, src_m, esize, num_elements, &result,
             [](uint64_t a, uint64_t b, uint8_t es) -> uint64_t {
-              int8_t shift = static_cast<int8_t>(b & 0xFF);
-              uint8_t bits = es * 8;
+              int32_t shift = static_cast<int8_t>(b & 0xFF);
+              uint32_t bits = es * 8;
               if (shift >= 0) {
-                return (shift >= bits) ? 0 : (a << shift);
+                return (static_cast<uint32_t>(shift) >= bits) ? 0 : (a << shift);
               } else {
-                int8_t rshift = -shift;
+                uint32_t rshift = static_cast<uint32_t>(-shift);
                 return (rshift >= bits) ? 0 : (a >> rshift);
               }
             });
